find changed samples from the point of divergence
instead of comparing the current commit to `main`, compare the current
commit to the commit from which we branched. (the common ancestor
between our branch and main). Otherwise, as main moves forward, our
changeset will include changes on our branch, and the inverse of the
changes on main (since those changes will not exist on our branch).